requirements.txt
streamlit
streamlit-option-menu
mysql-connector-python
pycryptodomex
pillow
numpy